backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
firebase-admin==6.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.9.2